requirements.txt
dotenv==0.9.9
imbalanced-learn==0.13.0
joblib==1.5.1
numpy==2.3.1
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.16.0
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.0.2
